Move get_or_create_remote_user method to UserManager
